rename, comment, and refactoring
