+++ skqulacs/circuit/__init__.py
@@ -1,9 +1,9 @@
 from .circuit import LearningCircuit
 from .graph import show_blochsphere
 from .pre_defined import (
-    create_deepqsv,
-    create_defqsv,
-    create_farhi_circuit,
-    create_farhi_watle,
+    create_shirai_ansatz,
+    create_ibm_embedding_circuit,
+    create_farhi_neven_ansatz,
+    create_farhi_neven_watle_ansatz,
     create_qcl_ansatz,
 )
