+++ skqulacs/qnn/classifier.py
@@ -7,8 +7,20 @@ from qulacs import Observable
 from scipy.optimize import minimize
 
 from skqulacs.circuit import LearningCircuit
-from skqulacs.qnn.qnnbase import QNN, _get_x_scale_param, _min_max_scaling
+from skqulacs.qnn.qnnbase import QNN
 from skqulacs.typing import Literal
+
+def _get_x_scale_param(x):
+    minimum = np.min(x, axis=0)
+    maximum = np.max(x, axis=0)
+    sa = (maximum - minimum) / 2
+    return [minimum, maximum, sa]
+
+
+def _min_max_scaling(x: List[List[float]], scale_x_param):
+    """[-1, 1]の範囲に規格化"""
+    # print([((xa - scale_x_param[0]) / scale_x_param[2]) - 1 for xa in x])
+    return [((xa - scale_x_param[0]) / scale_x_param[2]) - 1 for xa in x]
 
 
 class QNNClassifier(QNN):
