+++ tests/test_qnn_regressor_farhi.py
@@ -6,7 +6,7 @@ import pytest
 from numpy.random import default_rng
 from sklearn.metrics import mean_squared_error
 
-from skqulacs.circuit import create_farhi_circuit, create_farhi_watle
+from skqulacs.circuit import create_farhi_neven_ansatz, create_farhi_neven_watle_ansatz
 from skqulacs.qnn import QNNRegressor
 
 
@@ -41,9 +41,9 @@ def test_noisy_sine_two_vars(solver: str, maxiter: int, farhitype: str):
     depth = 6
 
     if farhitype == "watle":
-        circuit = create_farhi_watle(n_qubit, depth, 0)
+        circuit = create_farhi_neven_watle_ansatz(n_qubit, depth, 0)
     else:
-        circuit = create_farhi_circuit(n_qubit, depth, 0)
+        circuit = create_farhi_neven_ansatz(n_qubit, depth, 0)
 
     qnn = QNNRegressor(n_qubit, circuit, solver)
     qnn.fit(x_train, y_train, maxiter)
@@ -79,7 +79,7 @@ def test_noisy_sine(solver: str, maxiter: int):
 
     n_qubit = 4
     depth = 11
-    circuit = create_farhi_circuit(n_qubit, depth, 0)
+    circuit = create_farhi_neven_ansatz(n_qubit, depth, 0)
     qnn = QNNRegressor(n_qubit, circuit, solver)
     qnn.fit(x_train, y_train, maxiter)
     x_test, y_test = generate_noisy_sine(x_min, x_max, num_x)
